WIP: first pass, modelSpec schema for custom config
